+++ app/src/main/java/data/Avis.java
@@ -4,13 +4,13 @@ public class Avis {
     private String username;
     private String description;
     private String picture;
-    private int note;
+    private float note;
     private String restaurant;
 
     public Avis() {
     }
 
-    public Avis(String username, String description, String picture, int note, String restaurant) {
+    public Avis(String username, String description, String picture, float note, String restaurant) {
         this.username = username;
         this.description = description;
         this.picture = picture;
@@ -27,8 +27,8 @@ public class Avis {
     public String getPicture() { return picture; }
     public void setPicture(String picture) { this.picture = picture; }
 
-    public int getNote() { return note; }
-    public void setNote(int note) { this.note = note; }
+    public float getNote() { return note; }
+    public void setNote(float note) { this.note = note; }
 
     public String getRestaurant() { return restaurant; }
     public void setRestaurant(String restaurant) { this.restaurant = restaurant; }
